Trying the same thing as the last commit
Hopefully this works but I'll put the same message for reference.
This is the version that our Thunkable app was from the "Outline for TESTCASES.md" commit 4098c73. This allows us to keep all of our progress from the test cases and go back to the previous version of the app that works without having to use git reset.